Removed strike and sparge volumes from calculator
Changed logic in water calculator. Older system accepted strike and
sparge as fields. This made the API quite complex because most of
the operations work for both waters same way.

Now the calculator accepts water volume and usage, so instead of
using one calculator for both waters you are expected to create
instances for both waters